# public/openapi.yaml
openapi: 3.1.0
info:
  title: PdfPixels API
  description: |
    Free online image and PDF processing API. Compress, resize, convert, and edit images with professional-grade tools.
    
    ## Features
    - Image compression with quality or target size
    - Image resizing with multiple units (px, cm, inch)
    - Format conversion (JPG, PNG, WebP, AVIF, GIF, TIFF)
    - Image filters and effects
    - Rotation and flip
    - PDF merge, split (per page, range, or by file size), and compression
    - PDF to image and image to PDF conversion (JPG, PNG, WebP, TIFF, HEIC)
    - OCR for scanned PDFs
    - PDF text extraction, embedded image extraction, and text-to-PDF
    - Multi-size favicon (.ico) generation
    
    ## Authentication
    No authentication required. All endpoints are public.
    
    ## Rate Limits
    - 20 requests per minute per processing route per IP (pdf/*, image/*, ai)
    - 100 requests per minute global for other API routes
    - Contact (3/min) and newsletter (5/min) endpoints are stricter
    - Limits verified against production code on 2026-09-25; see https://www.pdfpixels.com/llms-full.txt
    
    ## Support
    Email: support@pdfpixels.com
    Website: https://www.pdfpixels.com
  version: 1.0.0
  contact:
    name: PdfPixels Support
    email: support@pdfpixels.com
    url: https://www.pdfpixels.com
  license:
    name: Free to Use
    url: https://www.pdfpixels.com/#terms
servers:
  - url: https://www.pdfpixels.com/api
...
paths:
...
  /pdf/split:
    post:
      summary: Split PDF
      description: Split a PDF into individual pages or extract specific pages
      operationId: splitPdf
      tags:
        - PDF Processing
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: PDF file to split (max 50MB). Split-all mode returns a ZIP truncated at 20 pages; range/single extract caps at 50 pages.
                mode:
                  type: string
                  description: "Split mode: 'all' returns one PDF per page (ZIP, truncated at 20 pages), 'range'/'single' extract a selection (max 50 pages), 'size' chunks the document into parts each under maxSizeMb (max 100 parts, 300 pages)."
                  enum: [all, range, single, size]
                  default: all
                pageRange:
                  type: string
                  description: Page range for extraction (e.g., "1-3,5,7-9")
                  example: "1-3,5"
                singlePage:
                  type: integer
                  description: Single page number to extract
                  example: 2
                maxSizeMb:
                  type: number
                  description: Size limit per part for 'size' mode (0.1-25 MB). Response is a ZIP of part-N.pdf files.
                  example: 25
      responses:
        '200':
          description: Successfully split PDF
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  mode:
                    type: string
                    example: split-all
                  totalPages:
                    type: integer
                    example: 10
                  pages:
                    type: array
                    description: Array of split pages (for 'all' mode)
                    items:
                      type: object
                      properties:
                        pageNumber:
                          type: integer
                        pdfUrl:
                          type: string
                          format: uri
                        fileName:
                          type: string
                  pdfUrl:
                    type: string
                    format: uri
                    description: Extracted PDF URL (for 'range' or 'single' mode)
                  extractedPages:
                    type: array
                    items:
                      type: integer
                    example: [1, 2, 3, 5]
...